+++ src/sdk/Dependencies/DependencyTypes.ts
@@ -1,26 +1,13 @@
-export interface LicensesList {
-    name: string;
-    spdxId: string;
-    isSpdxApproved: boolean;
+import { Dependency, DependencyFile, DependencyResponse } from "../Clients/Dependency/IDependencyClient";
+
+export interface IDependency extends Dependency {
+  scope?: string;
 }
 
-export interface DependenciesList {
-    component: string;
-    purl: string;
-    version?: string;
-    requirement?: string;
-    scope?: string;
-    licensesList: LicensesList[];
+export interface IDependencyFile extends Omit<DependencyFile, 'dependenciesList'> {
+  dependenciesList: IDependency[];
 }
 
-export interface FilesList {
-    file: string;
-    id: string;
-    status: string;
-    dependenciesList: DependenciesList[];
-}
-
-export interface IDependencyResponse {
-    filesList: FilesList[];
-    status?: string;
+export interface IDependencyResponse extends Omit<DependencyResponse, 'filesList'> {
+  filesList: IDependencyFile[];
 }