+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file. See [standa
 
 ## [Unreleased]
 - Upcoming changes...
-
-## [0.29.1] (2026-01-12)
 ### Fixed
 - Reset proxy environment variables on each new HttpClient and Dispatcher instance
+### Changed
+- Upgraded `form-data` dependency from ^4.0.0 to ^4.0.5
+- Upgraded `tar` dependency from ^6.1.11 to ^6.2.0
 
 ## [0.29.0] (2025-12-22)
 ### Changed
@@ -201,4 +202,3 @@ All notable changes to this project will be documented in this file. See [standa
 ### [0.28.0](https://github.com/scanoss/scanoss.js/compare/v0.27.0...v0.28.1) (2025-12-10)
 ### [0.28.1](https://github.com/scanoss/scanoss.js/compare/v0.28.0...v0.28.1) (2025-12-10)
 ### [0.29.0](https://github.com/scanoss/scanoss.js/compare/v0.28.1...v0.29.0) (2025-12-22)
-### [0.29.1](https://github.com/scanoss/scanoss.js/compare/v0.29.0...v0.29.1) (2026-01-12)