+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. See [standa
 ### Fixed
 - Reset proxy environment variables on each new HttpClient and Dispatcher instance
 - Include file path in decompression error messages
+- Added missing component URL to the HTML format
 ### Changed
 - Upgraded `form-data` dependency from ^4.0.0 to ^4.0.5
 - Upgraded `tar` dependency from ^6.1.11 to ^6.2.0