+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. See [standa
 ## [Unreleased]
 - Upcoming changes...
 
+## [0.30.0] (2026-01-14)
 ### Fixed
 - Reset proxy environment variables on each new HttpClient and Dispatcher instance
 - Include file path in decompression error messages
@@ -209,3 +210,5 @@ All notable changes to this project will be documented in this file. See [standa
 ### [0.28.0](https://github.com/scanoss/scanoss.js/compare/v0.27.0...v0.28.1) (2025-12-10)
 ### [0.28.1](https://github.com/scanoss/scanoss.js/compare/v0.28.0...v0.28.1) (2025-12-10)
 ### [0.29.0](https://github.com/scanoss/scanoss.js/compare/v0.28.1...v0.29.0) (2025-12-22)
+### [0.30.0](https://github.com/scanoss/scanoss.js/compare/v0.29.0...v0.30.0) (2026-01-14)
+